abstracts: add mock_area scaling factor for abstracts
Works by first doing floorplan, then doing a second run where
SYNTH_GUT=1 is set and the DIE/CORE_AREA is set to be a
mock_area multiple (larger or smaller) than the first floorplan
and generating an scaled abstract.

Then the regular abstract is generated, but the scaled .lef file
replaces the actual .lef file.

This yields a .lib file for mock-area case and a .lef file for
the mock-area case.

This allows separating concerns: set up a budget floorplan and
have timing information.

Lots of other updates to sharpen dependencies, such as only
depending on .lib files, not the whole abstract. This stops
synthesis from re-running if the .lib file is the same,
but the .lef file is changed to mock a smaller area.

Signed-off-by: Øyvind Harboe <oyvind.harboe@zylin.com>